+++ dd-java-agent/instrumentation/java-concurrent/src/main/java/datadog/trace/instrumentation/java/concurrent/CallableInstrumentation.java
@@ -1,0 +1,75 @@
+package datadog.trace.instrumentation.java.concurrent;
+
+import static datadog.trace.agent.tooling.bytebuddy.matcher.DDElementMatchers.implementsInterface;
+import static net.bytebuddy.matcher.ElementMatchers.isPublic;
+import static net.bytebuddy.matcher.ElementMatchers.named;
+import static net.bytebuddy.matcher.ElementMatchers.takesArguments;
+
+import com.google.auto.service.AutoService;
+import datadog.trace.agent.tooling.Instrumenter;
+import datadog.trace.bootstrap.ContextStore;
+import datadog.trace.bootstrap.InstrumentationContext;
+import datadog.trace.bootstrap.instrumentation.java.concurrent.State;
+import datadog.trace.context.TraceScope;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.concurrent.Callable;
+import lombok.extern.slf4j.Slf4j;
+import net.bytebuddy.asm.Advice;
+import net.bytebuddy.description.method.MethodDescription;
+import net.bytebuddy.description.type.TypeDescription;
+import net.bytebuddy.matcher.ElementMatcher;
+
+/** Instrument {@link Runnable} and {@Callable} */
+@Slf4j
+@AutoService(Instrumenter.class)
+public final class CallableInstrumentation extends Instrumenter.Default {
+
+  public CallableInstrumentation() {
+    super(AbstractExecutorInstrumentation.EXEC_NAME);
+  }
+
+  @Override
+  public ElementMatcher<TypeDescription> typeMatcher() {
+    return implementsInterface(named(Callable.class.getName()));
+  }
+
+  @Override
+  public String[] helperClassNames() {
+    return new String[] {
+      AdviceUtils.class.getName(),
+    };
+  }
+
+  @Override
+  public Map<String, String> contextStore() {
+    final Map<String, String> map = new HashMap<>();
+    map.put(Callable.class.getName(), State.class.getName());
+    return Collections.unmodifiableMap(map);
+  }
+
+  @Override
+  public Map<? extends ElementMatcher<? super MethodDescription>, String> transformers() {
+    final Map<ElementMatcher<? super MethodDescription>, String> transformers = new HashMap<>();
+    transformers.put(
+        named("call").and(takesArguments(0)).and(isPublic()),
+        CallableInstrumentation.class.getName() + "$CallableAdvice");
+    return transformers;
+  }
+
+  public static class CallableAdvice {
+
+    @Advice.OnMethodEnter(suppress = Throwable.class)
+    public static TraceScope enter(@Advice.This final Callable thiz) {
+      final ContextStore<Callable, State> contextStore =
+          InstrumentationContext.get(Callable.class, State.class);
+      return AdviceUtils.startTaskScope(contextStore, thiz);
+    }
+
+    @Advice.OnMethodExit(onThrowable = Throwable.class, suppress = Throwable.class)
+    public static void exit(@Advice.Enter final TraceScope scope) {
+      AdviceUtils.endTaskScope(scope);
+    }
+  }
+}
