+++ dd-java-agent/instrumentation/java-concurrent/src/main/java/datadog/trace/instrumentation/java/concurrent/RunnableInstrumentation.java
@@ -14,7 +14,6 @@ import datadog.trace.context.TraceScope;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
-import java.util.concurrent.Callable;
 import lombok.extern.slf4j.Slf4j;
 import net.bytebuddy.asm.Advice;
 import net.bytebuddy.description.method.MethodDescription;
@@ -24,15 +23,15 @@ import net.bytebuddy.matcher.ElementMatcher;
 /** Instrument {@link Runnable} and {@Callable} */
 @Slf4j
 @AutoService(Instrumenter.class)
-public final class RunnableCallableInstrumentation extends Instrumenter.Default {
+public final class RunnableInstrumentation extends Instrumenter.Default {
 
-  public RunnableCallableInstrumentation() {
+  public RunnableInstrumentation() {
     super(AbstractExecutorInstrumentation.EXEC_NAME);
   }
 
   @Override
   public ElementMatcher<TypeDescription> typeMatcher() {
-    return implementsInterface(named(Runnable.class.getName()).or(named(Callable.class.getName())));
+    return implementsInterface(named(Runnable.class.getName()));
   }
 
   @Override
@@ -46,7 +45,6 @@ public final class RunnableCallableInstrumentation extends Instrumenter.Default 
   public Map<String, String> contextStore() {
     final Map<String, String> map = new HashMap<>();
     map.put(Runnable.class.getName(), State.class.getName());
-    map.put(Callable.class.getName(), State.class.getName());
     return Collections.unmodifiableMap(map);
   }
 
@@ -55,10 +53,7 @@ public final class RunnableCallableInstrumentation extends Instrumenter.Default 
     final Map<ElementMatcher<? super MethodDescription>, String> transformers = new HashMap<>();
     transformers.put(
         named("run").and(takesArguments(0)).and(isPublic()),
-        RunnableCallableInstrumentation.class.getName() + "$RunnableAdvice");
-    transformers.put(
-        named("call").and(takesArguments(0)).and(isPublic()),
-        RunnableCallableInstrumentation.class.getName() + "$CallableAdvice");
+        RunnableInstrumentation.class.getName() + "$RunnableAdvice");
     return transformers;
   }
 
@@ -68,21 +63,6 @@ public final class RunnableCallableInstrumentation extends Instrumenter.Default 
     public static TraceScope enter(@Advice.This final Runnable thiz) {
       final ContextStore<Runnable, State> contextStore =
           InstrumentationContext.get(Runnable.class, State.class);
-      return AdviceUtils.startTaskScope(contextStore, thiz);
-    }
-
-    @Advice.OnMethodExit(onThrowable = Throwable.class, suppress = Throwable.class)
-    public static void exit(@Advice.Enter final TraceScope scope) {
-      AdviceUtils.endTaskScope(scope);
-    }
-  }
-
-  public static class CallableAdvice {
-
-    @Advice.OnMethodEnter(suppress = Throwable.class)
-    public static TraceScope enter(@Advice.This final Callable thiz) {
-      final ContextStore<Callable, State> contextStore =
-          InstrumentationContext.get(Callable.class, State.class);
       return AdviceUtils.startTaskScope(contextStore, thiz);
     }
